Make traversal evaluation tail-recursive
Formerly, the computation would be composed in reverse upon completion
of a traversal. Now, instead, we compose the computation as we go, in
a more iterative fashion. This makes it easier to keep track of the
computation as it follows the same order as the actual execution and
happens concurrently with it.