Revert to using lisp-specific (paredit) emit and capture for now
It would be nice to have a common (Symex) implementation of these
features for Lisp and tree-sitter, but it may be that doing these
operations purely in a structural way ends up having behavior (like
indenting according to the new structure) that we don't necessarily
want, as it often isn't invertible/idempotent. Handling trailing
whitespace also seems problematic and prone to special cases.
Understanding the right approach will take more careful analysis of
each case.

For now, we will revert to the earlier implementation as it
works well for Lisp (but doesn't for tree-sitter, of course), and
leave the tree-sitter features unimplemented. We can return to this in
the future to see what the best thing to do would be.