remove old call to `deactivate-mark`
I believe this is a leftover from when we used to use regions to
"highlight" symexes. Nowadays we use an overlay for this, so there's
no case I can think of (currently - in the future we may have
selection of regions via #53) where the region would be active while
in Symex mode.